pyproject.toml
[tool.poetry]
name = "skagit60-api"
version = "0.1.0"
description = "Backed for upgraded skagit60 site."
authors = ["raymond-devries <raymond.l.devries@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
motor = "^2.1.0"
uvicorn = "^0.11.8"
python-slugify = "^4.0.1"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = "^1.7.2"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-asyncio = "^0.14.0"
isort = "^5.3.2"
black = "^19.10b0"
pre-commit = "^2.6.0"
httpx = "^0.14.1"
factory-boy = "^2.12.0"
pytest-cov = "^2.10.1"
pytest-mock = "^3.2.0"

[tool.black]
target-version = ["py38"]
include = "app"
line_length = 88

[tool.isort]
src_paths = ["app"]
known_third_party = ["bson", "fastapi", "motor", "pydantic", "pytest", "uvicorn"]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
ensure_newline_before_comments = "True"
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
